move src/modules/index.js into src/main.js
